web_app/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
Flask==3.0.2
gunicorn==21.2.0
torch==2.2.1+cpu
torchvision==0.17.1+cpu
Pillow==10.2.0
